requirements.txt
kivy
pytesseract
pillow
treys
